Example of a tool which inspired the library
Takes a Reaper project file and a target directory. It then copies all of
the media files in that project to the target directory and creates a
new Reaper project with all of the file references changed to point in
this new target folder.

Is is useful when you mistakenly work with a Reaper project with the
wrong settings and your media files end up all over your disk. This
also happens quite easily when importing media files to a project.