Clarify state vs. mode because quoted vs default is the only state
I was anticipating other major states in the parsing than quote vs not
quoted. As of yet, this has been unnecessary. I decided to name the
state variable 'state' and the operating mode shall be called 'mode'
now on. I renamed 'idle' to 'default' and made sure I set the mode
similarly in all of the cases. Maybe I should use protocols or
something, but this seems to work with only 1 nested if & cond. If I
start to need more nesting I should probably look into some kind of
polymorphism with multimethods or protocols.